refactor: makefile added
