+++ README.md
@@ -1,9 +1,17 @@
-# [WIP]: HTMX-boilerplate
+# [WIP]: HTMX-todo
 
 This repository is dedicated to create a Copy/Paste style boiler plate for HTMX projects, this is the way I personally use it
 next step would be adding a command line tool to create the project template boiler plate
+
+## template
+
+the template uses an simple todoList application using sqlite3 as the database
 
 WIP:
 
 - [ ] Add Support for `.templ` files
 - [ ] Add command line way to auto generate all the files without git clone
+- [ ] Add http/2
+- [ ] Add websocket
+- [ ] Add authentication handler
+- [ ] Add database
