+++ internal/handler/layouts.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"fmt"
@@ -7,15 +7,15 @@ import (
 	"net/http"
 )
 
-type LayoutHandlers struct {
+type Layout struct {
 	tmpl     *template.Template
 	lg       *slog.Logger
 	siteName string
 	root     string
 }
 
-func NewLayoutHandler(tmp *template.Template, sn, rootTempName string, lg *slog.Logger) *LayoutHandlers {
-	return &LayoutHandlers{
+func NewLayout(tmp *template.Template, sn, rootTempName string, lg *slog.Logger) *Layout {
+	return &Layout{
 		tmpl:     tmp,
 		lg:       lg,
 		siteName: sn,
@@ -23,7 +23,7 @@ func NewLayoutHandler(tmp *template.Template, sn, rootTempName string, lg *slog.
 	}
 }
 
-func (lh *LayoutHandlers) BlogHandler(w http.ResponseWriter, r *http.Request) {
+func (lh *Layout) BlogHandler(w http.ResponseWriter, r *http.Request) {
 	data := LayoutInfo{
 		SubTmplName: "Blog",
 		PageName:    fmt.Sprintf("%s > Blog", lh.siteName),
@@ -32,7 +32,7 @@ func (lh *LayoutHandlers) BlogHandler(w http.ResponseWriter, r *http.Request) {
 	lh.renderTemplate(data, w, r)
 }
 
-func (lh *LayoutHandlers) IndexHandler(w http.ResponseWriter, r *http.Request) {
+func (lh *Layout) IndexHandler(w http.ResponseWriter, r *http.Request) {
 	data := LayoutInfo{
 		SubTmplName: "",
 		PageName:    lh.siteName,
@@ -41,7 +41,7 @@ func (lh *LayoutHandlers) IndexHandler(w http.ResponseWriter, r *http.Request) {
 	lh.renderTemplate(data, w, r)
 }
 
-func (lh *LayoutHandlers) AboutHandler(w http.ResponseWriter, r *http.Request) {
+func (lh *Layout) AboutHandler(w http.ResponseWriter, r *http.Request) {
 	data := LayoutInfo{
 		SubTmplName: "About",
 		PageName:    fmt.Sprintf("%s > About", lh.siteName),
@@ -50,7 +50,7 @@ func (lh *LayoutHandlers) AboutHandler(w http.ResponseWriter, r *http.Request) {
 	lh.renderTemplate(data, w, r)
 }
 
-func (lh *LayoutHandlers) TagsHandler(w http.ResponseWriter, r *http.Request) {
+func (lh *Layout) TagsHandler(w http.ResponseWriter, r *http.Request) {
 	data := LayoutInfo{
 		SubTmplName: "Tags",
 		PageName:    fmt.Sprintf("%s > Tags", lh.siteName),
@@ -59,7 +59,7 @@ func (lh *LayoutHandlers) TagsHandler(w http.ResponseWriter, r *http.Request) {
 	lh.renderTemplate(data, w, r)
 }
 
-func (lh *LayoutHandlers) renderTemplate(data LayoutInfo, w http.ResponseWriter, r *http.Request) {
+func (lh *Layout) renderTemplate(data LayoutInfo, w http.ResponseWriter, r *http.Request) {
 	w.Header().Add(contentType, textHtml)
 	err := lh.tmpl.ExecuteTemplate(w, lh.root, data)
 	if err != nil {
