+++ makefile
@@ -1,4 +1,4 @@
-APP_NAME = htmx
+APP_NAME = htmx-todo
 APP_VERSION := $(shell git describe --abbrev=0 || echo v0.1.0)
 GIT_HEAD := $(shell git rev-parse --verify HEAD)
 BUILD_AT := $(shell date --rfc-3339 'seconds' -u)
