+++ tailwind.config.js
@@ -2,12 +2,7 @@
 module.exports = {
   content: ["./public/**/*.{html,js}", "./templates/**/*.html"],
   theme: {
-    extend: {
-      // fontFamily:{
-      //   'display': ['Oswald'],
-      //   'body': ['"Open Sans"']
-      // }
-    },
+    extend: {},
   },
   plugins: [],
 };
